requirements.txt
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
colorama==0.4.1
cycler==0.10.0
imutils==0.5.2
isort==4.3.20
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.16.4
olefile==0.46
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
Pillow==6.1.0
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
typed-ast==1.3.5
wincertstore==0.2
wrapt==1.11.1
mkl-service==2.0.2
